FS-8389: [build] Fix msvc 2015 build warnings

+++ src/mod/applications/mod_conference/conference_video.c
@@ -1483,7 +1483,7 @@ void *SWITCH_THREAD_FUNC conference_video_muxing_write_thread_run(switch_thread_
 					switch_time_t now = switch_time_now();
 					
 					if (last) {
-						int delta = now - last;
+						int delta = (int)(now - last);
 						if (delta > member->conference->video_fps.ms * 5000) {
 							switch_core_session_request_video_refresh(member->session);							
 						}
